SLSCMN-3 move middleware tests
